refactor(post): extract MarkdownPostService from PostCommands
The /post markdown command carried ~290 lines of pure, collaborator-free logic -
HTML/horizontal-rule stripping, deep-heading folding, LaTeX-to-Unicode, Markdown
table -> fixed-width code block, and recursive length-aware section splitting -
inlined in the command, so none of it could be unit-tested.

Move it all into MarkdownPostService.render(String): List<String> (no Discord,
DB, config or network dependencies). PostCommands becomes a thin adapter that
downloads the attachment, calls render, and sends the sections. Behaviour is
unchanged.
